requirements.txt
openai>=2.0.0
Pillow>=12.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
colorama>=0.4.6
requests>=2.31.0
numpy>=2.0.0

